% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schooldataset.R
\docType{data}
\name{schooldataset}
\alias{schooldataset}
\title{Data for Student Levels in Qatar}
\format{
A tibble with 120 rows and 8 variables:
\describe{
  \item{Year}{chr that specify the year which academic year}
  \item{pass_fail}{chr that describe the overall grade if pass or fail}
  \item{level}{chr which describe the levels: Preparatory, Primary, Secandary, Sec-Specialzation}
  \item{grade}{chr grade 1-6}
  \item{gender}{chr female or male}
}
}
\source{
\url{https://www.data.gov.qa/explore/dataset/school-results/table/}
}
\usage{
schooldataset
}
\description{
This dataset contains total of pass and failed students, by level, grade and gender.
}
\keyword{datasets}
